requirements.txt
numpy>=1.21
scipy>=1.7
pandas>=1.3
scikit-learn>=1.0
matplotlib>=3.5
seaborn>=0.11
streamlit>=1.20

# Text processing (optional lemmatization)
nltk>=3.7

# PDF + OCR pipeline (optional but used when PDFs uploaded)
pdfplumber>=0.7
pdf2image>=1.16
pytesseract>=0.3.10
pillow>=9.0

# Reporting
fpdf2>=2.7

# Optional visualization
wordcloud>=1.9
